+++ services/preview/mysql/mgmt/2021-05-01-preview/mysqlflexibleservers/CHANGELOG.md
@@ -1,2 +1,26 @@
-# Change History
+# Unreleased
 
+## Breaking Changes
+
+### Signature Changes
+
+#### Funcs
+
+1. ServersClient.ListByResourceGroup
+	- Params
+		- From: context.Context, string
+		- To: context.Context, string, string
+1. ServersClient.ListByResourceGroupComplete
+	- Params
+		- From: context.Context, string
+		- To: context.Context, string, string
+1. ServersClient.ListByResourceGroupPreparer
+	- Params
+		- From: context.Context, string
+		- To: context.Context, string, string
+
+## Additive Changes
+
+### New Constants
+
+1. CreateMode.CreateModeMigrate